Move jira notificator to oar package
