+++ pyproject.toml
@@ -7,7 +7,7 @@ requires = [
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-packages = ["oar", "tools"]
+packages = ["oar"]
 
 [tool.setuptools_scm]
 write_to = "oar/_version.py"
